Rework usage of Token-generation
In the future, 'tasks' (i.e. StanfordNER, FOX) which modify the tokens are seperated from the processing-pipeline. Instead, they are implementing an interface 'TaskInterface' and the processing-pipeline can collect these tasks and execute them on the tasks one after the other.

StanfordNER: Instead of using a webclient, it is no integrated into the application itself.