Add n-gram hierarchy from SESSA
These classes provide a way to split the question into multiple n-grams.
As we not only want to search for unigrams, but also bigger n-grams (think e.g. "Angela Merkel"), we need a way to build a hierarchy to filter out unnecessary n-grams (e.g. we may find results for "Angela" and "Angela Merkel", but ultimately only want the bigger one